+++ eslint.config.js
@@ -23,10 +23,7 @@ export default tseslint.config(
     },
     rules: {
       ...reactHooks.configs.recommended.rules,
-      "react-refresh/only-export-components": [
-        "warn",
-        { allowConstantExport: true },
-      ],
+      "react-refresh/only-export-components": ["warn"],
       "@typescript-eslint/no-unused-vars": [
         "error",
         {
