+++ src/components/admin/form/form-item-context.ts
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+export type FormItemContextValue = {
+  id: string;
+  name: string;
+};
+
+export const FormItemContext = createContext<FormItemContextValue>(
+  {} as FormItemContextValue,
+);
