+++ src/components/rich-text-input/default-editor-options.ts
@@ -1,0 +1,3 @@
+import type { UseEditorOptions } from "@tiptap/react";
+
+export const DefaultEditorOptions: Partial<UseEditorOptions> = {};
