+++ src/components/rich-text-input/index.ts
@@ -1,2 +1,4 @@
 export * from "./rich-text-input";
 export * from "./rich-text-input-toolbar";
+export * from "./use-rich-text-input-editor";
+export * from "./default-editor-options";
