+++ src/components/rich-text-input/use-rich-text-input-editor.ts
@@ -1,0 +1,7 @@
+import { useTiptapEditor } from "./minimal-tiptap/hooks/use-tiptap-editor";
+
+export const useRichTextInputEditor = () => {
+  const { editor } = useTiptapEditor();
+
+  return editor;
+};
