+++ website/src/components/ui/button.tsx
@@ -56,5 +56,4 @@ function Button({
   )
 }
 
-// eslint-disable-next-line react-refresh/only-export-components
 export { Button, buttonVariants }
