+++ website/src/components/ui/navigation-menu.tsx
@@ -164,6 +164,5 @@ export {
   NavigationMenuLink,
   NavigationMenuIndicator,
   NavigationMenuViewport,
-  // eslint-disable-next-line react-refresh/only-export-components
   navigationMenuTriggerStyle,
 }
